Added checkout and seeder

+++ sakenny/API/Controllers/ServicesController.cs
@@ -48,7 +48,7 @@ namespace sakenny.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete(long id)
+        public async Task<IActionResult> Delete(int id)
         {
             try
             {
